src/requirements/local.txt
-r base.txt

# Development tools
django-debug-toolbar==1.6